syntax change to uint

+++ contracts/DiamondFacet.sol
@@ -15,18 +15,18 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
     function cut(DiamondCut[] memory _diamondCuts) public override {         
         require(msg.sender == $contractOwner, "Must own the contract.");        
         // loop through diamond cuts
-        for(uint256 diamondCutIndex; diamondCutIndex < _diamondCuts.length; diamondCutIndex++) {
+        for(uint diamondCutIndex; diamondCutIndex < _diamondCuts.length; diamondCutIndex++) {
             DiamondCut memory diamondCut = _diamondCuts[diamondCutIndex];            
             // loop through facet cuts
-            for(uint256 facetCutIndex; facetCutIndex < diamondCut.facetCuts.length; facetCutIndex++) {
+            for(uint facetCutIndex; facetCutIndex < diamondCut.facetCuts.length; facetCutIndex++) {
                 FacetCut memory facetCut = diamondCut.facetCuts[facetCutIndex];
                 // removing functions                
                 if(facetCut.action == CutAction.Remove) {                    
-                    for(uint256 i; i < facetCut.functionSelectors.length; i++) {
+                    for(uint i; i < facetCut.functionSelectors.length; i++) {
                         bytes4 selector = facetCut.functionSelectors[i];
                         require($facets[selector] != address(0), "Function does not exist. Can't remove.");
-                        uint256 index = $funcSelectorToIndex[selector];                        
-                        uint256 lastIndex = $funcSelectors.length - 1;
+                        uint index = $funcSelectorToIndex[selector];                        
+                        uint lastIndex = $funcSelectors.length - 1;
                         if (index != lastIndex) {
                             $funcSelectors[index] = $funcSelectors[lastIndex];
                             $funcSelectorToIndex[$funcSelectors[lastIndex]] = index;
@@ -38,7 +38,7 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
                 }
                 else {                    
                     // Check the size of the facet contract to ensure it is a contract.
-                    uint256 contractSize;
+                    uint contractSize;
                     address facet = facetCut.facet;
                     assembly {
                         contractSize := extcodesize(facet)
@@ -46,7 +46,7 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
                     require(contractSize > 0, "Facet address is not a contract and is not address(0)");
                     // replace functions
                     if(facetCut.action == CutAction.Replace) {
-                        for(uint256 i; i < facetCut.functionSelectors.length; i++) {
+                        for(uint i; i < facetCut.functionSelectors.length; i++) {
                             bytes4 selector = facetCut.functionSelectors[i]; 
                             require($facets[selector] != address(0), "Function does not exist. Can't replace.");
                             $facets[selector] = facetCut.facet;                            
@@ -54,7 +54,7 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
                     }
                     // add functions
                     else if(facetCut.action == CutAction.Add) {                        
-                        for(uint256 i = 0; i < facetCut.functionSelectors.length; i++) {
+                        for(uint i = 0; i < facetCut.functionSelectors.length; i++) {
                             bytes4 selector = facetCut.functionSelectors[i];
                             require($facets[selector] == address(0), "Function already exists. Can't add.");
                             $funcSelectorToIndex[selector] = $funcSelectors.length;
@@ -71,11 +71,11 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
         emit DiamondCuts(_diamondCuts);
     }
 
-    function totalFunctions() external view override returns(uint256) {
+    function totalFunctions() external view override returns(uint) {
         return $funcSelectors.length;
     }
 
-    function functionSelectorByIndex(uint256 _index) external view override returns(bytes4 functionSelector, address facet) {
+    function functionSelectorByIndex(uint _index) external view override returns(bytes4 functionSelector, address facet) {
         require(_index < $funcSelectors.length, "Function index does not exist.");
         functionSelector = $funcSelectors[_index];
         facet = $facets[functionSelector];
@@ -83,27 +83,27 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
     
     function facets() external view override returns(Facet[] memory) {
         // get default size of arrays
-        uint256 defaultSize = $funcSelectors.length;        
+        uint defaultSize = $funcSelectors.length;        
         if(defaultSize > 20) {
             defaultSize = 20;
         }
         Facet[] memory facets_ = new Facet[](defaultSize);
         uint8[] memory numFacetSelectors = new uint8[](defaultSize);
-        uint256 numFacets;
+        uint numFacets;
         // loop through function selectors
-        for(uint256 selectorsIndex; selectorsIndex < $funcSelectors.length; selectorsIndex++) {
+        for(uint selectorsIndex; selectorsIndex < $funcSelectors.length; selectorsIndex++) {
             bytes4 selector = $funcSelectors[selectorsIndex];
             address facet = $facets[selector];
             bool continueLoop = false;     
             // loop through collected facets
-            for(uint256 facetIndex; facetIndex < numFacets; facetIndex++) {
+            for(uint facetIndex; facetIndex < numFacets; facetIndex++) {
                 if(facets_[facetIndex].facet == facet) {                    
-                    uint256 arrayLength = facets_[facetIndex].functionSelectors.length;
+                    uint arrayLength = facets_[facetIndex].functionSelectors.length;
                     // if array is too small then enlarge it
                     if(numFacetSelectors[facetIndex]+1 > arrayLength) {
                         bytes4[] memory biggerArray = new bytes4[](arrayLength + defaultSize);
                         // copy contents of old array
-                        for(uint256 i; i < arrayLength; i++) {
+                        for(uint i; i < arrayLength; i++) {
                             biggerArray[i] = facets_[facetIndex].functionSelectors[i];
                         }
                         facets_[facetIndex].functionSelectors = biggerArray;
@@ -120,12 +120,12 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
                 continueLoop = false;
                 continue;
             }
-            uint256 arrayLength = facets_.length;
+            uint arrayLength = facets_.length;
             // if array is too small then enlarge it
             if(numFacets+1 > arrayLength) {
                 Facet[] memory biggerArray = new Facet[](arrayLength + defaultSize);
                 uint8[] memory biggerArray2 = new uint8[](arrayLength + defaultSize);
-                for(uint256 i; i < arrayLength; i++) {
+                for(uint i; i < arrayLength; i++) {
                     biggerArray[i] = facets_[i];
                     biggerArray2[i] = numFacetSelectors[i];        
                 }
@@ -138,12 +138,12 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
             numFacetSelectors[numFacets] = 1;
             numFacets++;
         }
-        uint256 difference = facets_.length - numFacets;
+        uint difference = facets_.length - numFacets;
         // shorten the array
         assembly {
             mstore(facets_, sub(mload(facets_), difference))
         }
-        for(uint256 facetIndex; facetIndex < numFacets; facetIndex++) {
+        for(uint facetIndex; facetIndex < numFacets; facetIndex++) {
             bytes4[] memory selectors = facets_[facetIndex].functionSelectors;
             difference = selectors.length - numFacetSelectors[facetIndex];
             // shorten the array
@@ -155,10 +155,10 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
     }
    
     function facetFunctionSelectors(address _facet) external view override returns(bytes4[] memory) {
-        uint256 funcSelectorsLength = $funcSelectors.length;
-        uint256 numFacetSelectors;        
+        uint funcSelectorsLength = $funcSelectors.length;
+        uint numFacetSelectors;        
         bytes4[] memory facetSelectors = new bytes4[](funcSelectorsLength);        
-        for(uint256 selectorsIndex; selectorsIndex < funcSelectorsLength; selectorsIndex++) {
+        for(uint selectorsIndex; selectorsIndex < funcSelectorsLength; selectorsIndex++) {
             bytes4 selector = $funcSelectors[selectorsIndex];            
             if(_facet == $facets[selector]) {
                 facetSelectors[numFacetSelectors] = selector;          
@@ -166,7 +166,7 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
             }
         }
         // shorten array
-        uint256 difference = funcSelectorsLength - numFacetSelectors;
+        uint difference = funcSelectorsLength - numFacetSelectors;
         assembly {
             mstore(facetSelectors, sub(mload(facetSelectors), difference))
         }
@@ -178,13 +178,13 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
     }
 
     function facetAddresses() external view override returns(address[] memory) {
-        uint256 funcSelectorsLength = $funcSelectors.length;
+        uint funcSelectorsLength = $funcSelectors.length;
         address[] memory facets_ = new address[](funcSelectorsLength);
-        uint256 numFacets;        
-         for(uint256 selectorsIndex; selectorsIndex < funcSelectorsLength; selectorsIndex++) {
+        uint numFacets;        
+         for(uint selectorsIndex; selectorsIndex < funcSelectorsLength; selectorsIndex++) {
             address facet = $facets[$funcSelectors[selectorsIndex]]; 
             bool continueLoop = false;
-            for(uint256 facetIndex; facetIndex < numFacets; facetIndex++) {
+            for(uint facetIndex; facetIndex < numFacets; facetIndex++) {
                 if(facet == facets_[facetIndex]) {
                     continueLoop = true;
                     break;
@@ -198,7 +198,7 @@ contract DiamondFacet is Diamond, DiamondLoupe, Storage {
             numFacets++;            
         }
         // shorten array
-        uint256 difference = funcSelectorsLength - numFacets;
+        uint difference = funcSelectorsLength - numFacets;
         assembly {
             mstore(facets_, sub(mload(facets_), difference))
         }
